Back/requirements.txt
adjustText==1.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.6.3
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
fonttools==4.53.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib-metadata==6.8.0
importlib_resources==6.4.0
install-jdk==1.1.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.7.1
joblib==1.4.2
jpype1==1.5.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
konlpy==0.6.0
locust==2.17.0
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
msgpack==1.0.7
nest-asyncio==1.6.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.55.0
packaging==24.1
pandas==2.1.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.6
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.4.0
roundrobin==0.0.4
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
timedelta==2020.12.3
tomli==2.0.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
